refactor?(api): move dependency injection to learner management bounded context
